requirements.txt
path
pillow
git+https://github.com/jevinskie/pytrofs#egg=pytrofs
